refactor : 기본 테스트 위치 변경
