+++ src/test/java/kitchenpos/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.Acceptance;
+package kitchenpos;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
